refactor: Change API service code directory structure
API calls are used across tabs, they are shared rather than limited to
specific page.
