+++ packages/frontend/src/apis/workoutsAPI.ts
@@ -1,11 +1,5 @@
-import { IExercise,IRoutine, IWorkout } from './types';
-
-const MINUTE_MILLISECS = 60 * 1000;
-export const DAY_MILLISECS = 24 * 60 * MINUTE_MILLISECS;
-
-export const getDateString = (date: Date):string => new Date(date.getTime() - date.getTimezoneOffset() * MINUTE_MILLISECS).toISOString().split('T')[0];
-
-export const getTodayString = ():string => getDateString(new Date());
+import { IExercise, IRoutine, IWorkout } from '../Daily/types';
+import { getDateString } from '../utils/getDateString';
 
 export const getWorkouts = async (date: Date):Promise<IWorkout[]> => {
   const dateString = getDateString(date);
