+++ packages/frontend/src/hooks/useExercises.ts
@@ -1,7 +1,7 @@
 import { useQuery, UseQueryResult } from "react-query";
 
+import { getExercises } from "../apis/exercisesAPI";
 import { IExercise } from "../Daily/types";
-import { getExercises } from "../Exercises/services";
 
 export function useExercises(): UseQueryResult<IExercise[], Error> {
   return useQuery<IExercise[], Error>(`exercises`, () => getExercises());
